minor changes in file DEL

+++ src/events/handlers/file.rs
@@ -30,7 +30,7 @@ pub async fn put(
     // Serialize and validate
     let file_input = <PubkyAppFile as Validatable>::try_from(&blob, &file_id)?;
 
-    //debug!("file input {:?}", file_input);
+    debug!("file input {:?}", file_input);
 
     let file_meta = ingest(&user_id, file_id.as_str(), &file_input).await?;
 
@@ -58,21 +58,20 @@ pub async fn put(
 async fn ingest(
     user_id: &PubkyId,
     file_id: &str,
-    pubkyapp_file: &PubkyAppFile,
-    //client: &PubkyClient,
+    pubkyapp_file: &PubkyAppFile
 ) -> Result<FileMeta, DynError> {
     let pubky_client = PubkyConnector::get_pubky_client()?;
+
     let blob = match pubky_client.get(pubkyapp_file.src.as_str()).await? {
         Some(metadata) => metadata,
+        // TODO: Shape the error to avoid the retyManager
         None => return Err("EVENT ERROR: no metadata in the file blob".into()),
     };
-
-    //debug!("File Metadata: {:?}\n{:?}", file_id, blob);
+    
     store_blob(file_id.to_string(), user_id.to_string(), &blob).await?;
 
-    let static_path = format!("{}/{}", user_id, file_id);
     Ok(FileMeta {
-        urls: FileUrls { main: static_path },
+        urls: FileUrls { main: format!("{}/{}", user_id, file_id) },
     })
 }
 
@@ -115,13 +114,14 @@ pub async fn del(user_id: &PubkyId, file_id: String) -> Result<(), DynError> {
     )
     .await?;
 
-    let file = &result[0];
+    if !result.is_empty() {
+        let file = &result[0];
 
-    if let Some(value) = file {
-        value.delete().await?;
+        if let Some(value) = file {
+            value.delete().await?;
+        }
+        remove_blob(file_id, user_id.to_string()).await?;
     }
-
-    remove_blob(file_id, user_id.to_string()).await?;
 
     Ok(())
 }
